Introduce {Set,Clear,Complement,Test}BitVector functions
There are functions which operate on 32bit and 64bit variables, but
their inline assembly is not suitable for bigger arrays without forcing
variable they operate on always to memory, which is not desirable.
Existing functions also do not report previous value of the bit.

So let's roll out versions that can operate on bit arrays.  As all I
need is 256 bits, I've implemented only version with 32bit index.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>