Introduce SOCK_SEQPACKET to vSockets.
Finally, the mythical sequential packet socket type appears.  This socket
type is specifically for guestRPC, over which Tools can communicate with
the VMX.  It is *not* a general sequential socket type: while you can
certainly make these sockets on the host, they won't be able to connect
anywhere, and you can't open such sockets inside the VMX, you can only
receive connections from the guest.

It works by using DGRAMs underneath.  vSocket DGRAMs are bit-compatible
with VMCI DGRAMs, so we can get from vSockets in the guest directly to the
VMX, without having to go through the host vSockets driver.

To make it connected and reliable, we introduce a very simple protocol
using a new packet format.  It consists only of CONNECT, DATA, SHUTDOWN and
CLOSE (the latter two being mostly the same as far as the VMX is
concerned).  Datagrams going to the VMX are already reliable.  In the
other direction, if we fail to receive, we send a CLOSE down with an error
(unlike the STREAM protocol, I actually remembered to embed an error code
this time).  For DATA packets, we place payload after the packet header.
We have to copy in from user-mode anyway, so it doesn't add any copy
overhead.

This is part 1, which is the Linux guest side.  It adds a new socket type
to our family and implements the appropriate socket functions (things like
accept() and so forth are just unecessary).

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>