Fix time synchronization when restarting tools
When the tools are restarted, and the clock is modified (out of sync
with the host's clock), if 'timeSync' feature is enabled, then the clock
is not synced. This happens only first time after the tools are
restarted. There is no issue from the second time onwards.

In the code, we do the one step synchronization only if the tool status
is stopped. Ideally, we should do this one step synchronization the first
time even after the tools are restarted. This is fixed by setting the
tools status to stopped when the tools are restarted initially.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>