Don't use $HOME to get the current user's home directory on Mac OS
There are some cases on Mac OS where the $HOME environment variable has
not been set even though the current user does have a valid home
directory. For example, if you enable the "root" user in the UI and then
try to start Fusion via Finder, it would fail spectacularly because it
was unable to figure out the Log and Preferences directories.

In lib/misc, for Mac OS, instead of resorting to $HOME in the case where
we are trying to expand "~" for the current user, we instead use the
posix passwd API, which according to the man page is properly hooked up
to DirectoryServices and thus will give us any user's homedir even if it
isn't actually in the /etc/passwd file.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>