Cleanup and rearchitecture of HGFS server change notification part VII
Final initial change set for fixing FSEvents polling for Lion OS X 10.7.
This is currently an OS X host only file, but now contains the main logic
to handle the file system change notifications and is platform independent.
The OS X platform specific code is now in hgfsFSEventsMac files.

In a later change set this file will be renamed to reflect that it is
common to all platforms while the platform dependent parts will implement
the FSEvent API (as done for OS X in part VI).

Currently only the OS X platform supports the change notification feature
and as such only the scons build file has been changed.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>