GuestInfo: ignore the extra NIC information
If the total number of network interfaces configured for a VM exceeds
the maximum XDR limit, then the guestInfo plugin crashes.

GuestInfo plugin calls intf_loop() to retrieve the information about all
the available interfaces. ReadInterfaceDetails() is passed as a callback
function to intf_loop(). Add a check for the number of NICs processed
and ignore the extra NIC information.  By doing this, VMware Tools
returns the maximum information that it could return and drops all the
extra NIC info on the floor.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>