Fix toolbox 'shrink' feature on newer distributions using bind mounts
VMware toolbox-cmd utility provides two disk related features i.e.
shrink and wipe. When the user executes 'shrink' or 'wipe' command, the
utility parses /etc/mtab and checks if there is an entry matching the
specific mount point. If a match exists, then the utlity immediately
stops the parsing. This approach is not very robust and may cause few
issues especially if there are multiple entries for the specified mount
point. If the first entry maps to an unsupported partition type and the
second entry maps to a supported partition type, the current parsing
logic returns EX_UNAVAILABLE error. We should fix this.

Modified the code to continue parsing the entries if the first match entry
maps to an unsupported partition type.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>