Fix error on file deletion on read-only HGFS share
The failure was being done at the set file information call of the
Windows client instead of the create call. This only occurred when
communicating to HGFS servers on Posix platforms. The Posix platform
only uses the open mode of the HGFS open protocol request even if the
desired access is valid. On Windows HGFS servers the desired access is
checked and used if valid, and if not falls back to the Posix open mode
call.

However, for the case of delete requests, the Windows client does not
set up the Posix open mode quite correctly. It igonores the DELETE
desired access flag.  This should be mapped to a Posix open mode
requesting write access. So for occasions where read attributes and
delete is requested in the desired access the Posix open mode is still
passed as read-only. This causes the open to succeed on a read-only
share when it should be failed with access denied.

This fix is to address the HGFS Posix server to try to map a Windows
client passing the desired access mask to the Posix open mode will then
allow all Windows clients to have the open for delete access validated
by the HGFS server.  Linux clients are unaffected and do not use this
field for the open requests.  It also has the benefit of making the
Posix and Windows HGFS servers more compatible.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>