VMBLOCK: fix panic on unmounting on FreeBSD 9.0
FreeBSD added a new operation vop_advlockpurge that is called
from vgonel() after an inode was released. We should not attempt
to pass this operation down to the underlying filesystem and
instead install vop_stdadvlockpurge() as our handler, exactly like
nullfs does.

Although there wasn't explicit bump of __FreeBSD_version for this
change, 900013 is the closest one and we'll use it to detect when
when we need to install vop_advlockpurge handler.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>