Fix HGFS file append on OS X
File appends were incorrect on OS X as they appear to be handled
differently by posix pwrite calls from Linux. On Linux if the handle is
opened for append mode then pwrite will ignore the offset and append to
the file. This is not correct for OS X, where the offset is still used.

For OS X we now check if the handle is opened for append and use the
write call instead.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>