Fix Lion guest and host files on shared folder show diff contents
The OS X hgfs client reference counts file handles that are shared for
accessing file data on the HGFS server. Unfortunately the implementor
failed to actually verify if the reference counting worked by testing
even the simplest sequence of file IO operations: opening a file,
reading it, and close should result in the reference count being set to
zero and the close request was sent to the HGFS server.  This did not
occur, as a result the open handle remainded forever open on the server
until the HGFS server was forced to shutdown. This open handle was
reused for any further accesses from the client until it was rebooted.
The behavior of OS X (and Linux) file systems mean that the HGFS client
will only receive the data that it has written to that file or the
original data stored in the file which is now cached. If a host
application modifies the contents of that file the HGFS client will not
see it, until the old handle gets closed and a new one opened.

The fix is for the reference count for the open handle to be correctly
handled.  This will mean that the matching close for an open handle will
reduce the reference count to zero and therefore a closse request will
be sent to the HGFS server. The next access to the file will get the
latest file data on disk from the host.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>