Add an entry for 'oracle linux' in distroArray.
The 'guestinfo' plugin in VMware Tools is responsible for retrieving
details about the guest (like guest family, guest id, guest fullname
etc.). The 'guestinfo' plugin uses HostinfoOSData to retrieve the data
about the guest OS. HostinfoOSData() uses 'lsb_release' package to
reteive the details. If the 'lsb_release' command fails for some reason,
then /etc/*-release files are parsed to retrieve the necessary
information. There is a specific list of /etc/*-release files that are
parsed.

We don't have an entry for 'Oracle Linux'. We need to add an entry for
'/etc/oracle-release' file in the list. In 'Oracle Linux' guests, both
files '/etc/redhat_release' and /etc/oracle_release' exist. So, we need
to add the new entry in the code before the 'redhat' entry. Otherwise,
'/etc/redhat-release' is parsed first and improper information is
retrieved.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>