vSock: only print warning if we could not register address family at all
vSockets on Linux tries to claim an address family starting from the top
of the valid range.  If someone has already claimed an address, we will
print a warning to the kernel log and then continue looping.  This
warning started appearing on Cent OS 5.4 in autoinstall tests, presumably
because someone is grabbing the top family.  Because the automation is
screenshot-based, the presence of the warning is causing the test to fail.
Fixed by moving the warning to the very end.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>