Handle special files in HGFileCopy library.
Certain files like /proc/kallsyms can be copied from the guest only if
the special 'sequential open' mode is specified. This changeset removes
that constraint. Implemented the support in HGFilecopy library to
specify the sequential mode flag automatically if it thinks the remote
file is a special file.

This is done by implementing the support in hgfsServer to specify the
file as 'sequential only' if it thinks so. HGFSServer sets this bit when
the attributes are queried for the file. Implemented a new HGFile copy
option which a user can specify if he wants to automatically handle
the special 'sequential only' files. If the new file copy option is set,
then the HGFilecopy library automatically sets the 'sequential open' flag
if it thinks the file being copied is a 'sequential only' file.

Made changes to the vix library code to use the new HGFileCopy option
as default for the file copy functions.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>