VMCI/vSOCK: fix event soft lockup on Linux
We make the silly mistake of assuming in our QP event CBs that we are in BH
context, because we didn't ask for a delayed CB.  But in this case, VMCI
doesn't make any guarantees about what context we are in, it could be BH or
process, we don't know.  Fixed so that we don't make this silly assumption.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>