vmmemctl: use capability based system to advertise balloon features
Use a capability based system instead of versioning to advertise
and negotiate balloon features.

On START command, the guest is sending its capability and the
monitor returns the masked capability. That way, the guest knows
exactly what features are supported by the monitor.

Currently the capability use two bits: if the bit 1 is set, then
the unbatched lock and unlock are supported, and if the bit 2 is
set, then the batched lock and unlock are supported. Bit 1 was
chosen to be backward compatible with "protocol v2", and bit 0
is unused to avoid issue with the screwed "protocol v3".

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>