Fix directory listing of stale information in OS X guests
The directory readdir uses the existing handle of the directory continually
even if restarting the listing of entries. This leads to stale data and if the
handle is kept open for long periods due to open handles of files within that
folder (parental references) the user will never see uptodate information.

This fix is the simplest fix for now. We simply refresh the folder handle on
any new readdir query. This closes the existing handle and issues a new handle
open.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>