HGFS: clean up HGFS lock type
The HGFS protocol has duplicated the lock types from previous versions
with later implementations. This is the first of a couple of changes to tidy
this part of the code up and later remove unncessary duplication of types.

This change renames the previous one to better reflect what it represents.
The previous type name indicated a lock not really its type.

No HGFS server has shipped with the locking protocol implemented and so
to maintain correctness for backwards compatibility there are two issues:
- all lock fields should remain the same size
- the default value indicating no lock retains its value of zero

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>