Remove the limit of 64 operating systems in the guest OS table
The guest OS table is currently limited to 64 entries because each
guest OS is stored as a bit in a uint64 and bitmask operations are
used to determine whether a guest belongs to a particular set.
This change moves to a list of OS ID's, so there is no limit on the
total number of guest operating systems.  There is still a limit of
64 on the total number of guests that can be a member of a set, such
as ALLLINUX or ALLDARWIN, because the implementation requires a fixed
size array for the list of guests in a set, but the limit is arbitrary.
To get this all to work, the vmmovhds scheme script needed to change
from bit checking to list processing, and the perl script that
generates the scheme include file needed some help.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>