HGFS: extract out the oplock specific code into their own files
The main HGFS server code and file is cluttered and has been used as a
general dumping ground for abosultely everything to do with the HGFS
server.  This has resulted in files that have become way too large and hard
to navigate.  This change is to help reduce that by moving the oplock code
into their own set of files. This helps platforms that do not have support
for this feature from having to code contain too much extra logic. Also for
those platforms that do support this complex feature, they can hide all
that complexity away.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>