HGFS: Cleanup of server platform specific functions to follow convention
The HgfsServerScandir function does not follow the naming convention of the
HGFS server and platform specific code. Therefore, to prevent confusion
about the implementation of the function being platform specific it is
renamed to be HgfsPlatformScandir.

Also the platform initialization and destruction names did not follow the
convention so modified those too.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>