HGFS: Fix server directory read hack for the virtual shares folder
For some uncommented reason, the HGFS virtual shares folder was always
refreshed on the first search read even if this was the first search read
since the search open. This means that the initial scan of shares on the
search open was always thrown away - making it a total waste of time even
constructing the list of dents. Secondly, the was an incorrect assumption
that the search type was always the shares folder if there was not a share
path. This maybe true for Posix (Linux/OS X) hosts but certainly not true
for Windows.

Replace this crap.

Fixed the function to call into the server platform specific function to
handle refreshing directory listings and handle the search type
appropriately according to what is supported on the platform.

Track whether the directory listing has been already returned completely to
the client and use that to determine whether a restart (and rescan of the
contents) is required. This means we don't scan the virtual directory
contents twice for the first read request from the client.

Renamed HgfsServerGetDents to match more closely what it really does and
make it consistent with the equivalent function name for operating on real
folders.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>