HGFS: Fix Windows server virtual folders
For the UNC or DRIVE virtual folders enabled by setting the share's host
path to an empty string, currently are not quite handled correctly.
The shares with empty host names are listed in their Shared Folders virtual
folder but have incorrect attriubtes. This is because the shares host paths
are passed to the real get attributes function which fails when the path
is an empty string. These entries should be defaulted to the HGFS virtual
folder attributes.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>