HGFS: Cleanup the directory entry code to be really platform specific - part I
The directory entry work is very inefficient and will cause oplock issues.
For Windows platforms the directory reads will change and thus the assumptions
about the platform specific DirectoryEntry structure will not hold true.
Therefore, the DirectoryEntry code is moving to the platform specific files
and will allow the code to have complete control over the structure and contents.
Initially, there will be some duplication of code but that will change in later
work.
There is already bugs with this code and places where ifdefs are added for
platform specific cases not handled by some platforms. These should not be
in platform independent code.
The volume query is also broken and misses some of the Windows platform specific
cases completely which should be handled and will be in the Windows only code.
(Again, later change sets.)

For now, this just simply moves a routine out of the common code into the platform
specific code. More changes to follow.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>