HGFS: Cleanup the directory entry code to be really platform specific - part II
The directory entry work is very inefficient and will cause oplock issues.
For Windows platforms the directory reads will change and thus the assumptions
about the platform specific DirectoryEntry structure will not hold true.
Therefore, the DirectoryEntry code is moving to the platform specific files
and will allow the code to have complete control over the structure and contents.
Initially, there will be some duplication of code but that will change in later
work.
There is already bugs with this code and places where ifdefs are added for
platform specific cases not handled by some platforms. These should not be
in platform independent code.
The volume query is also broken and misses some of the Windows platform specific
cases completely which should be handled and will be in the Windows only code.
(Again, later change sets.)

For now, this just simply moves two routines out of the common code into the platform
specific code. The routines have some minimal cleanup to remove the ifdefs or remove
the ifdefs and code not required for each of the platforms.

More changes to follow for moving the volume query handler code and adding the
missing virtual folder cases in Windows.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>