Add guest OS definitions for Linux 3.X
Originally I thought we could do this by aliasing Linux 3.x to Linux
2.6.x, but that only works inbound.  Once the guest is identified and
aliased to 2.6.x, outbound messages and buttons will incorrectly
identify the guest as 2.6.x instead of 3.x, which is sloppy and will be
confusing to customers.  Instead, this change adds gos table entries for
Linux 3.x, which uses the last two (of 64 max) entries in the table.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>