HGFS: Improve efficiency of volume queries on a virtual folder
For some virtual folders like the one that contains all the shares,
volume query size information necessitates an enumeration of all the
shares. This allows the query to determine which particular volume
information to return, since the shares can reside on many different
volumes on the host. The algorithm will pick that largest or the
smallest volume size to return depending on the configuration settings.

When enumerating the logic currently gets and removes the first entry
from shares list. This results in the table of shares being shuffled up
for each removal. If there are a lot of shares this results in a lot of
memmove calls and the larger the table the more memory is moved each
time. Since the list of entries are stored in an array, it would be
simpler to just remove the last entry each time not the first. The order
of the enumeration is irrelevant.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>