HGFS: Posix server fix an incorrect size passed for name conversion
The directory read records may contain directory entries with names
that are of UTF8 form D and are converted to form C which the clients
will expect (e.g. Windows).
The name and size passed was assumed to be the whole name field of the
directory entry structure. However, the record length is the value of
the valid data in that record which can be a lot less than the total
size of the name field. The copied records only have the requisite
record valid data size allocated and not the whole complete structure.
This should at least optimize that call and stop any possibility of the
conversion routine going beyond the valid record.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>