HGFS: Cleanup the directory entry code to be really platform specific - part V
Moving the platform dependent definitions to the platform specific files.
They should be private types to each platform so each can define and use
them as efficiently as makes sense for that platform.

Now moving the structure types to the platform only code. Only a forward
declaration is now public.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>