HGFS: Fix server channel characteristics for HGFS sessions
Currently the HGFS session had hardcoded constants specific to VMCI
set in the HGFS session object and which is returned to the client on a
HGFS create session protocol request. The constant was not used by any
client that was released but will be in the future releases.
The HGFS session also can be used with the backdoor channel too, so
this first change is to move the constant (maximum packet size) from
the HGFS server into the channel specific code which passes it to the
HGFS server on a connect call. The HGFS server connect call will create
the HGFS server transport session object which will be used when creating
the HGFS session object.

Later changes will deal with the backward compatibility issues by updating
the HGFS session protocol request to allow newer clients that use the session
field for maximum packet size to determine the version of the server and
correctness of the session reply field.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>