HGFS: session creation server cleanup for channel characteristics
The HGFS server now returns the correct packet size for the session
protocol reply for the transport in use. This previously was always
the VMCI setting. Now we need to allow the clients to differentiate
the setting is valid for the transport and is not hardcoded to VMCI.
In order to achieve this, the 64 bit reserved field is split into
two 32 bit fields: flags and reserved.

New clients will be able to detect if the server is using just a hard
coded value or has indeed set it per transport characteristics.
Older existing clients just ignore the packet size field.

Made the session bool for change notification into a flag which is
also returned to the client. This is an additional check that can be
made that the server supports change notification other
than are the individual HGFS opcode requests being supported.

Fixed an issue that prevents an errant client sending remove watch opcodes to the server from being processed
when notification is not active.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>