Make Str_Snprintf_C_locale work again
It looks like previous change has broken the C_locale friendly print
function, namely it changes things back to using the localeconv() table.
This seems to have been done in the name of making android work,
however, it breaks our ability to correctly generate shaders on systems
when the locale uses "," to denote a decimal point.  This change reverts
that bit of behaviour so that the bsd_vsnprintf function uses the
provided decimal point unless we're on android.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>