HGFS: Improve the Windows server directory reads
The current implementation of the HGFS server directory read handling is
not efficient and results in lots of system calls for each entry returned
in the list. This will result in poor performance and in particular breaking
open oplocks. Both of these are unnecessary if we use the Windows native API.

So first instead of using the DirectoryEntry structure that makes no sense for
Windows and only filling in the name from each Win32 FindFirst, FindNext call,
we now use the same structure as the native API. Secondly, using the native API
we make one call with a sufficiently large buffer to read a large number of entries
per system call not just one at a time. Finally, the native API returns all the
required information for any record returned to the client so we don't have to
iterate through the entire list and do a GetFileAttributes on each. (Hence, no
open and query, close - which will break any open oplocks.)

Also now we get the directory entry record filled at the search open call when
we scan the folder and get all the entries. Previously, this was only the name
and the GetFileAttributes was returned for each at the search read time.
This change is also done for the virtual folders e.g. the shares under
"\\vmware-host\Shared Folders" and for any share where the host path is set to
empty thus showing the UNC and DRIVE virtual folders.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>