HGFS: add the oplock break protocol request packet handler in the server
When the HGFS server receives an oplock break from the file system callback
it will send the HGFS protocol oplock break request on to the client to handle
and send back an acknowledge.

This code just adds the packet handling for the protocol oplock break request
to the HGFS server. Since oplocks are not yet enabled this code is not yet
used but will be with subsequent changes.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>