Clean up the compat_vmtruncate
As of kernel 3.8 the vmtruncate call was removed. Also it should be noted
that tools for kernels prior to 2.6.9 have been frozen. Therefore, the
compat_vmtruncate macro is now only defined for tools for kernels 2.6.9
and upto and including kernel 3.7.

It is not expected to be used by any tools on kernel 3.8 and newer and if
so they should use the memory manager calls directly that vmtruncate called.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>