One copy of lib/err globals
It looks like we have multiple copies of errNumTable but only one
errStrTable. Each new copy of errNumTable comes for each library that
includes lib/err. Since errStrTable is not declared static, only one
exists between all shared libraries.

This change makes both errNumTable and errPtrTable non-static as
well. The intention is that we'll have only one copy of the errno
accounting across the entire binary.

Generally sharing globals across a bunch of libraries sounds
dangerous, but in this case it's an atomic hashtable so it works.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>