HGFS: start cleanup of packet handlers and add oplock break acknowledge
When the client receives a oplock break request from the HGFS server it
must reply to the Hgfs server with an oplock break acknowledge or close
the handle.

The oplock break acknowlege reply is now defined more completely and it
now has a packet handler.

The packet handlers now stop overriding the input packet const and really
do treat it as read only. This involved adding const to the request packet
handlers for the unpacking of client requests. This is only the first part
of dealing with this issue. There was some minor spillage into the HgfsServer
mainline code, which will be additionally cleaned up later. Further changes
will also include the platform specific request handlers too.

This now means that the server will not modify the incoming request data and
should be less prone to accidental overwrites, and also better document the
fact that the file names returned are in fact, just pointers into the incoming
packet and not duplicated strings. Hence, the mainline code does not attempt
to free any of the cross-platform names that are extracted from incoming
requests.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>