HGFS: Fix performance issues for Linux client
Revalidation of a file's attributes occurs way too often.

Unfortunately the fix revolves around the host ID for the file in
question which is incorrectly retrieved from Windows HGFS server. This
results in the file ID for the file always being zero. Hence, it is
repeatedly revalidated by getting the file attributes from the server.
The file ID is never updated and set due to the wrong flag being tested.

The file ID is always retrieved using the same common utility routine
and so all the places that get a file or folders attributes use it.
Hence, checking in the one place fixes about 4 different call sites -
root dirent, regular dirent and files.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>