HGFS: Cleanup of server request packets handlers, part II
Propagate the treatment of the input request data as constant. This means
finish off the cpName usage to use const. Remove casts and make the open
and directory info data structures use const too.

Corrected the name of HgfsServerGetShareInfo to be more meaningful to what
it really is doing, since the input name argument type to be const which
is how it was used. Made the function static and removed the declaration
from the header file.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>