HGFS: fix up the Hgfs header flags and information fields
Currently, the Hgfs server just blindly sets the header information
field to the host platform's internal error value.  This isn't very
handy as it will involve determining whether it is Posix or Win32.  Then
mapping accordingly to the Windows client or Linux client internal
kernel error values.

It would be more suited to being used in conjunction with a flag to
state what the field contains, so that it can be used for multiple
things if required - albeit not simultaneously. This would also allow
the client to know that the field is really valid, and not some host
specific error value which is currently set. Also, since the clients
only require one or two host error extended information, it probably is
simpler to define a crossplatform value for that new error value and
have the clients map it as they already do for the existing
crossplatform errors.

Also there was a flag for the flags field defined but never used. So it
is being thrown away and redefined. The flag was previously defined to
be a notification bug was intended to be more generic meaning host
server initiated request and not a reply to a client initiated request.
Fixed this.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>