VMCI/VSOCK: Pt1. Switch to upstreamed Linux API
This is the first part of a change to switch to the upstreamed Linux
API.  This change just flips the API names to match the upstreamed
versions.  A subsequent change will take care of fixing up the
parameters (we removed/changed some for upstreaming).

This change updates: VMCI, vSockets, the VMCI kernel lib on Windows,
the VMCI test driver, HGFS, dvFilter and vPageChannel.

It does NOT affect the user-mode VMCI API: there's no point, and it'd
be a waste of time changing all the tests.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>