HGFS: cleanup server request and header unpacking part I
The Hgfs server HgfsParseRequest routine to unpack packet header and
validation is somewhat tangled and convoluted. This attempts to clean
that up and make way for checking for the header flags if available
and any other information that maybe passed from the client on a per
request basis in the future.

The changes include:
- rename the HgfsParseRequest to HgfsUnpackPacketParams which is more
correct and consistent with the naming scheme for HGFS unmarshalling of
arguments.
- Split out the unpacking of the different protocol versions of header
and packet details into their respective subroutines thus simplifying
the main routine itself.
- Split out the HGFS session lookup and/or creation that occurs as part
of this routine into a subroutine call.

This cleans up some ordering issues of packet validation and data
extraction from that packet. It helps clean up some of the session
lookup which should be handled separately from this function which is
server specific and not protocol packet definition. A later change set
to deal with that.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>