HGFS: cleanup server request and header unpacking part II
After the first cleanup and extracting the header details by individual
routines and now we have the session information and routines separated.

1) Cleanup the unpack params routine:
- Make it just do the unpacking of the headers and validate the fields.
- Make any bad header sizes or incorrect values of the header fiels untrusted
data and we drop the packet by returning internal error. It makes no sense
to try and reply if the data in the header is not valid.
- Move the transport/server Hgfs packet handling, the session info and the
input params back to the hgfs server main file into a parent function of the
unpack params. The new parent function HgfsServerGetRequest is called from
the server receive message routine now instead.
- Make the input request handling variables const as the input data is not
modified.

2) Cleanup the processing a little more in the new HgfsServerGetRequest
function by doing the can fail processing first: Hgfs Packet, and unpack
packet params first. If we have valid data, determine the session and
then allocate and set the input params. Saves a potential processing
with the input params not initialized to anything but zero.

3) Remove the useless and duplicated function HgfsValidatePacket.

4) Move HgfsPackReplyHeaderV4 function declaration to the correct header file.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>