Link the VMCI socket common code into the VMX.
Use the common VMCI socket driver code in the VMX to provide access to
raw VMCI sockets. This functionality has some limitations, for example
blocking calls are not possible because blocking a VMX thread can
prevent the guest from making the necessary progress and cause deadlock.

A new lock parameter has been added to VSockSocket_Socket to allow the
caller to provide a lock for use by the socket. This is similar to the
lock provided to AsyncSocket and allows an AsyncSocket and VSockSocket
to share the same lock.

VMCI_DatagramSend requires a needsLock parameter. This is currently
hardcoded to FALSE in all cases. This may not be correct and needs to be
analyzed.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>