HGFS: add new flags to Hgfs create session request
The Hgfs create session request assumes currently that the only flags
required are on a per request type basis. This is simply not true. There
are some features that trancend many operations and will require a
general per session flag to denote support.

Two examples that immediately come to mind: oplocks and short name
support.

For oplocks the server is going to need to know ahead of any requests
that the client is going to be using this feature if supported by the
server. Furthermore, oplocks affect the whole operation of requests as
far as the server is concerned.  The server should handle requests
asynchronously if oplocks are in play to prevent deadlock. Also, oplocks
are taken out a file open time but can be broken by many other requests
not just concurrent opens.

Also for short name support for the host, these can occur not just with
directory listings and enumeration but for the client opens too.
Currently, the client must know if the server does support short names
so that it does not try and look up the short name itself.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>